fix: split CLI transport files
