+++ crates/dcc-mcp-cli/src/presentation/cli.rs
@@ -11,7 +11,7 @@ use super::output::{ExitCode, OutputFormat, OutputWriter, failure_envelope, to_j
 use crate::application::call_attribution::{
     attach_agent_session_id, attach_batch_agent_session_id,
 };
-use crate::application::client::{DccMcpClient, TransportMode};
+use crate::application::client::DccMcpClient;
 use crate::application::control_plane::DccControlPlane;
 use crate::application::doctor::{DoctorContext, run_doctor};
 use crate::application::gateway_ctrl;
@@ -33,6 +33,7 @@ mod job_progress;
 mod lint;
 mod marketplace_output;
 mod record_replay;
+mod transport_args;
 mod ui_control_output;
 
 #[cfg(test)]
@@ -41,6 +42,7 @@ use image_artifacts::{default_image_artifact_root, materialize_call_images};
 use job_progress::JobProgressReporter;
 use marketplace_output::reload_marketplace_value;
 use record_replay::{RecordReplayAction, run_record_replay};
+use transport_args::TransportArgs;
 use ui_control_output::compact_ui_control_result;
 
 use super::feedback_cmd::FeedbackArgs;
@@ -99,15 +101,8 @@ pub struct Args {
     /// Global timeout in seconds for all operations.
     #[arg(long, global = true, env = "DCC_MCP_TIMEOUT_SECS")]
     timeout_secs: Option<u64>,
-    /// Tool-call transport: auto (REST, then MCP fallback), rest, or mcp.
-    #[arg(
-        long,
-        global = true,
-        env = "DCC_MCP_CLI_TRANSPORT",
-        value_enum,
-        default_value_t = TransportMode::Auto
-    )]
-    transport: TransportMode,
+    #[command(flatten)]
+    transport: TransportArgs,
     #[command(subcommand)]
     command: Command,
 }
@@ -582,10 +577,11 @@ async fn run_with_args(args: Args) -> anyhow::Result<()> {
         output,
         non_interactive,
         timeout_secs: global_timeout_secs,
-        transport,
+        transport: transport_args,
         command,
     } = args;
 
+    let transport = transport_args.transport;
     let output = match &command {
         Command::Feedback(args) => args.resolve_output(output),
         Command::Install { json: true, .. } => Ok(OutputFormat::Json),
@@ -1476,7 +1472,6 @@ fn endpoint_for_mcp(raw: &str) -> String {
     }
 }
 
-/// Check whether a command timeout would be ignored in favor of a distinct global timeout.
 fn command_has_distinct_per_timeout(command: &Command, global_timeout_secs: Option<u64>) -> bool {
     let per_command_timeout = match command {
         Command::Smoke { timeout_secs, .. }
